requirements.txt
altgraph==0.17.4
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.8
imageio==2.37.0
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.0
networkx==3.4.2
numpy==2.2.2
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.0
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.16.1
PyQt6==6.8.0
PyQt6-Qt6==6.8.1
PyQt6_sip==13.9.1
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-image==0.25.1
scipy==1.15.1
setuptools==75.8.0
six==1.17.0
tifffile==2025.1.10
tzdata==2025.1
